finance_calendar/pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"

requires = [
  "setuptools",
  "setuptools-scm",
]

[project]
name = "fincalendar"
requires-python = ">=3.12"
classifiers = [
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.12",
]
dynamic = [
  "version",
]
dependencies = [
  "pydantic",
  "python-dateutil",
  "tabulate",
]

optional-dependencies.dev = [
  "mypy",
  "pyproject-fmt",
  "ruff",
  "types-python-dateutil",
  "types-tabulate",
]
scripts.fincalendar = "fincalendar.main:main"

[tool.ruff]
format.skip-magic-trailing-comma = true
lint.select = [
  "E4",
  "E7",
  "E9",
  "F",
  "I",
]
lint.isort.split-on-trailing-comma = false
